+++ tests/testthat/test_other.R
@@ -4,15 +4,18 @@ library(recipes)
 data(okc)
 
 set.seed(19)
-in_train <- sample(1:nrow(okc), size = 30000)
+in_test <- sample(1:nrow(okc), size = 200)
 
-okc_tr <- okc[ in_train,]
-okc_te <- okc[-in_train,]
+okc_tr <- okc[-in_test,]
+okc_te <- okc[ in_test,]
 
 rec <- recipe(~ diet + location, data = okc_tr)
 
+# assume no novel levels here but test later:
+# all(sort(unique(okc_tr$location)) == sort(unique(okc$location)))
+
 test_that('default inputs', {
-  others <- rec %>% step_other(diet, location)
+  others <- rec %>% step_other(diet, location, other = "another")
 
   tidy_exp_un <- tibble(
     terms = c("diet", "location"),
@@ -31,7 +34,6 @@ test_that('default inputs', {
       "oakland", "san francisco")
   )
   expect_equal(tidy_exp_tr, tidy(others, number = 1))
-
 
   diet_props <- table(okc_tr$diet)/sum(!is.na(okc_tr$diet))
   diet_props <- sort(diet_props, decreasing = TRUE)
@@ -91,15 +93,15 @@ test_that('high threshold - much removals', {
 
 
 test_that('low threshold - no removals', {
-  others <- rec %>% step_other(diet, location, threshold = 10^-10)
+  others <- rec %>% step_other(diet, location, threshold = 10^-30, other = "another")
   others <- prep(others, training = okc_tr, stringsAsFactors = FALSE)
   others_te <- bake(others, newdata = okc_te)
 
-  expect_equal(others$steps[[1]]$objects$diet$collapse, FALSE)
-  expect_equal(others$steps[[1]]$objects$location$collapse, FALSE)
-
-  expect_equal(okc_te$diet, others_te$diet)
-  expect_equal(okc_te$location, others_te$location)
+  expect_equal(is.na(okc_te$diet), is.na(others_te$diet))
+  expect_equal(is.na(okc_te$location), is.na(others_te$location))
+  
+  expect_equal(okc_te$diet, as.character(others_te$diet))
+  expect_equal(okc_te$location, as.character(others_te$location))
 })
 
 
@@ -108,8 +110,8 @@ test_that('factor inputs', {
   okc$diet <- as.factor(okc$diet)
   okc$location <- as.factor(okc$location)
 
-  okc_tr <- okc[ in_train,]
-  okc_te <- okc[-in_train,]
+  okc_tr <- okc[-in_test,]
+  okc_te <- okc[ in_test,]
 
   rec <- recipe(~ diet + location, data = okc_tr)
 
@@ -143,6 +145,36 @@ test_that('factor inputs', {
   expect_equal(is.na(okc_te$location), is.na(others_te$location))
 })
 
+
+test_that('novel levels', {
+  df <- data.frame(
+    y = c(1,0,1,1,0,0,0,1,1,1,0,0,1,0,1,0,0,0,1,0),
+    x1 = c('A','B','B','B','B','A','A','A','B','A','A','B',
+           'A','C','C','B','A','B','C','A'),
+    stringsAsFactors = FALSE)
+  training <- df[1:10,]
+  testing <- df[11:20,]
+  training$y <- as.factor(training$y)
+  training$x1 <- as.factor(training$x1)
+  testing$y <- as.factor(testing$y)
+  testing$x1 <- as.factor(testing$x1)
+  
+  novel_level <- recipe(y ~ ., data = training) %>%
+    step_other(x1) 
+  
+  novel_level <- prep(novel_level, training = training, retain = TRUE)
+  new_results <- bake(novel_level, newdata = testing)
+  orig_results <- bake(novel_level, newdata = training)
+  expect_true(all(new_results$x1[testing$x1 == "C"] == "other"))
+  expect_true(!any(orig_results$x1 == "other"))  
+})
+
+test_that("'other' already in use", {
+  others <- rec %>% step_other(diet, location, threshold = 10^-10)
+  expect_error(
+    prep(others, training = okc_tr, stringsAsFactors = FALSE)
+  )
+})
 
 test_that('printing', {
   rec <- rec %>% step_other(diet, location)